+++ system/sync.js
@@ -1,0 +1,40 @@
+const sql = require("../plugins/sql");
+const valid = require("../plugins/checkvalid");
+const logger = require("../plugins/logger");
+
+/**
+ * 
+ * @param {sql} sqlPlugin 
+ * @param {logger} log 
+ * @param {*} req 
+ * @param {*} res 
+ */
+module.exports = (sqlPlugin,log,mailer,req,res)=>{
+    /**
+     * @type {object}
+     */
+    const dataReceived = req.body;
+
+    const account = dataReceived["account"];
+    const cookie = dataReceived["cookie"];
+    const year = dataReceived["year"];
+    const user = dataReceived["user"];
+
+    if(!valid(dataReceived,["account","cookie","user","year"])){
+        res.sendStatus(400);
+        return;
+    }
+    
+    let ret = sqlPlugin.checkHash(account,cookie);
+    if (ret==null||ret["accountType"]=="empolyee"){
+        res.sendStatus(403);
+        return;
+    }
+    
+    sqlPlugin.syncTickets(user,year);
+
+    res.json({
+        "status":200
+    });
+  
+}